# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_api_handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  robot_scheduler_internal_model_dto.PCDFileCreateRequest:
    properties:
      area:
        description: 区域描述
        type: string
      extraInfo:
        description: 扩展信息
        type: string
      minioPath:
        description: MinIO存储路径
        type: string
      name:
        description: 地图名称
        maxLength: 100
        minLength: 1
        type: string
      path:
        description: 文件存储路径
        type: string
      size:
        description: 文件大小
        minimum: 0
        type: integer
      userName:
        description: 上传人员
        type: string
    required:
    - area
    - name
    - path
    - size
    - userName
    type: object
  robot_scheduler_internal_model_dto.PCDFileUpdateRequest:
    properties:
      area:
        description: 区域描述
        type: string
      extraInfo:
        description: 扩展信息
        type: string
      minioPath:
        description: MinIO存储路径
        type: string
      name:
        description: 地图名称
        maxLength: 100
        minLength: 1
        type: string
      path:
        description: 文件存储路径
        type: string
      size:
        description: 文件大小
        minimum: 0
        type: integer
      userName:
        description: 上传人员
        type: string
    type: object
  robot_scheduler_internal_model_dto.PCDFileUploadTokenRequest:
    properties:
      fileName:
        description: 原始文件名
        type: string
      size:
        description: 文件大小(字节)
        minimum: 0
        type: integer
    required:
    - fileName
    - size
    type: object
  robot_scheduler_internal_model_dto.SemanticMapCreateRequest:
    properties:
      extraInfo:
        description: 扩展信息
        type: string
      pcdFileId:
        description: 对应的pcd地图文件id
        type: integer
      semanticInfo:
        description: 语义信息
        type: string
      userName:
        description: 编辑人员
        type: string
    required:
    - pcdFileId
    - semanticInfo
    - userName
    type: object
  robot_scheduler_internal_model_dto.SemanticMapUpdateRequest:
    properties:
      extraInfo:
        description: 扩展信息
        type: string
      pcdFileId:
        description: 对应的pcd地图文件id
        type: integer
      semanticInfo:
        description: 语义信息
        type: string
      userName:
        description: 编辑人员
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 机器人调度系统后台管理接口
  termsOfService: http://swagger.io/terms/
  title: 机器人调度系统 API
  version: "1.0"
paths:
...
  /maps/pcd-files:
    get:
      consumes:
      - application/json
      description: 查询所有点云地图
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 查询点云地图列表
      tags:
      - 点云地图
    post:
      consumes:
      - application/json
      description: 创建新点云地图
      parameters:
      - description: 点云地图信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.PCDFileCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 创建点云地图
      tags:
      - 点云地图
  /maps/pcd-files/{id}:
    delete:
      consumes:
      - application/json
      description: 删除点云地图（软删除）
      parameters:
      - description: 点云地图ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 删除点云地图
      tags:
      - 点云地图
    get:
      consumes:
      - application/json
      description: 根据ID获取点云地图信息
      parameters:
      - description: 点云地图ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 点云地图不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取点云地图
      tags:
      - 点云地图
    put:
      consumes:
      - application/json
      description: 更新点云地图信息
      parameters:
      - description: 点云地图ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.PCDFileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 点云地图不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 更新点云地图
      tags:
      - 点云地图
  /maps/pcd-files/upload-token:
    post:
      consumes:
      - application/json
      description: 返回 MinIO 预签名上传 URL
      parameters:
      - description: 上传文件信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.PCDFileUploadTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取点云地图上传凭证
      tags:
      - 点云地图
  /maps/semantic-maps:
    get:
      consumes:
      - application/json
      description: 查询所有语义地图
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 查询语义地图列表
      tags:
      - 语义地图
    post:
      consumes:
      - application/json
      description: 创建新语义地图
      parameters:
      - description: 语义地图信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.SemanticMapCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 创建语义地图
      tags:
      - 语义地图
  /maps/semantic-maps/{id}:
    delete:
      consumes:
      - application/json
      description: 删除语义地图（软删除）
      parameters:
      - description: 语义地图ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 删除语义地图
      tags:
      - 语义地图
    get:
      consumes:
      - application/json
      description: 根据ID获取语义地图信息
      parameters:
      - description: 语义地图ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 语义地图不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取语义地图
      tags:
      - 语义地图
    put:
      consumes:
      - application/json
      description: 更新语义地图信息
      parameters:
      - description: 语义地图ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.SemanticMapUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 语义地图不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 更新语义地图
      tags:
      - 语义地图
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"